GuessingGame
This program can be used to check whether a required element is there in the huge list of catalog. Suppose we want to search for item in catalog, that contains 356789453 items, this game will take not more than 29 guesses to check whether the required item is present in the catalog.

The algorithm we have used here is binary search which takes (logn) to find an element in any sorted list.