MinSubArrVal
Given a startIndex and array, the function indexOfMinimum will find the minium value's position in the subarray(startIndex,array.length).
If we execute this in a loop, array will be sorted by finding minimum element and swapping it to the right poistion, which is the logic of Selection sort. Time complexity is O(n2).